requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
Django==3.1.3
django-extensions==3.0.9
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
passlib==1.7.4
password==0.2
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
urllib3==1.26.2
python-dotenv
